requirements.txt
django
djangorestframework
django-extensions
django-rest-multiple-models